Making Repository.delete method optional on value

+++ MJSwiftCore/Classes/Common/Repository/MappedRepository.swift
@@ -60,8 +60,11 @@ public class MappedRepository <From,To>: Repository <From> {
         return repository.putAll(toToMapper.map(array), in: query, operation: operation).map { self.toFromMapper.map($0) }
     }
     
-    public override func delete(_ value: From, in query: Query, operation: Operation) -> Future<Bool> {
-        return repository.delete(toToMapper.map(value), in: query, operation: operation)
+    public override func delete(_ value: From?, in query: Query, operation: Operation) -> Future<Bool> {
+        if let value = value {
+            return repository.delete(toToMapper.map(value), in: query, operation: operation)
+        }
+        return repository.delete(nil, in: query, operation: operation)
     }
     
     public override func deleteAll(_ array: [From], in query: Query, operation: Operation) -> Future<Bool> {
